Various Improvements, mainly around hooks (#65)
* Added afterUpdateQuery / afterInsertQuery

* Implement and document how to prevent default operations on active record

* remove outdated test (since it's no longer on readme)

* Added some really nice advanced topics.

* allow hasOne()->addField('foo'); (instead of saying foo twice)

* Applied fixes from StyleCI

* typos

* typo

+++ src/Model.php
@@ -564,6 +564,10 @@ class Model implements \ArrayAccess, \IteratorAggregate
             $this->unload();
         }
 
+        if ($this->hook('beforeLoad', [$id]) === false) {
+            return $this;
+        }
+
         $this->data = $this->persistence->load($this, $id);
         $this->id = $id;
         $this->hook('afterLoad');
@@ -739,7 +743,9 @@ class Model implements \ArrayAccess, \IteratorAggregate
                 }
             }
 
-            $this->hook('beforeInsert', [&$data]);
+            if ($this->hook('beforeInsert', [&$data]) === false) {
+                return $this;
+            }
 
             // Collect all data of a new record
             $this->id = $this->persistence->insert($this, $data);
@@ -854,7 +860,9 @@ class Model implements \ArrayAccess, \IteratorAggregate
 
             return $this;
         } elseif ($this->loaded()) {
-            $this->hook('beforeDelete', [$this->id]);
+            if ($this->hook('beforeDelete', [$this->id]) === false) {
+                return $this;
+            }
             $this->persistence->delete($this, $this->id);
             $this->hook('afterDelete', [$this->id]);
             $this->unload();
